+++ brewfiles.d/media.rb
@@ -5,6 +5,7 @@ brew 'flac'
 brew 'sox'
 brew 'libsndfile'
 brew 'atomicparsley'
+brew 'mozjpeg' # beets cover art encoding, see config.d/beets/plugins/bestjpeg.py
 
 cask 'foobar2000'
 cask 'iina'